MVP delivered - news summary fetching Guardian headlines and listing them with urls to Guardian website
